Apply cookiecutter template (#52)
* cruft link

* apply cookiecutter, get GHA template (horray!)

* replace coredev with crypto-eng, other minor things

* remove stuff related to core team project board

* remove circle ci

* apply a james suggestion to expand matrix

* markdown lint

* appease markdown linter

* more readme markdown lints

* more readme fixups

* more readme lints

* fixup my dependabot.yml

* fix dependabot yml for real this time

* yaml lint in ci.yaml

* increase nightly version

* permissions for installing gsl

* bump version of mc-crypto-rand we are pulling in

* address clippies

* remove duplicated dependabot file

* lint

* remove unrecognized "skip-children" option

* add license to test-helper crate

* make cargo deny happy

* fix rustflags in ci.yaml clobbering the correct rustflags

* add comment to the deny.toml

* Update .github/settings.yml


* Update .github/workflows/ci.yaml


---------

+++ mc-oblivious-map/src/lib.rs
@@ -591,7 +591,7 @@ where
                     oram.access(evicted_from, |block| {
                         let pairs: &mut [A8Bytes<Sum<KeySize, ValueSize>>] =
                             block.as_mut_aligned_chunks();
-                        let pair = &mut pairs[evicted_index as usize];
+                        let pair = &mut pairs[evicted_index];
                         let (key, val): (&mut A8Bytes<KeySize>, &mut A8Bytes<ValueSize>) =
                             pair.split();
